requirements.txt
flask==3.1.3
pillow==12.2.0
numpy==2.4.4
